Add support for include() in remote files; remove  need for proxy.
Download all files directly (rather than tunneling through a proxy
server) to remove dependence on running a special server.

On top of that change make the library and include base paths separate
so that when working with a remote file we can fetch included files with
the same base path of the original.

Finally, ensure that subdirectories in include paths work with the
gMemFs by truncating the path to the base name when checking for a
match.